Fixed IP proxy issues
When connecting through a proxy the request ip is changed to that of the
proxy. The application now checks the "X-Forwarded-For" header to see if
proxying both happened and was acknowledged. Necessary so that all
requests through a local load balancing or redirection proxy did not
fool the application into thinking that all connections were being made
by localhost.